vcpkg support
This PR lays the groundwork for vcpkg integration. The parts of the code that are meant to be executed only when packaged via vcpkg are put behind `ASMJIT_EXTERNAL` flag. The support is implemented in a manner that closely mirrors asmjit's support. I have tested this configuration with asmtk port in local vcpkg repo, and it worked flawlessly. Once those changes are merged, I will open a PR in vcpkg repo with portfiles for asmtk.

This PR would be instrumental in closing #25 and https://github.com/microsoft/vcpkg/issues/24359